- refactor : 회원 서비스 공통 로직 이동으로 인한 리팩터링
