+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/admin/AdminController.java
@@ -10,6 +10,7 @@ import ProjectDoge.StudentSoup.entity.school.Department;
 import ProjectDoge.StudentSoup.entity.school.School;
 import ProjectDoge.StudentSoup.repository.member.MemberRepository;
 import ProjectDoge.StudentSoup.service.DepartmentService;
+import ProjectDoge.StudentSoup.service.member.MemberCommonService;
 import ProjectDoge.StudentSoup.service.member.MemberService;
 import ProjectDoge.StudentSoup.service.SchoolService;
 import lombok.RequiredArgsConstructor;
@@ -29,6 +30,7 @@ import java.util.stream.Collectors;
 public class AdminController {
     private final MemberRepository memberRepository;
     private final MemberService memberService;
+    private final MemberCommonService memberCommonService;
     private final SchoolService schoolService;
     private final DepartmentService departmentService;
 
@@ -75,7 +77,7 @@ public class AdminController {
         log.info("MultipartFile : [{}]", updateForm.getMultipartFile());
         Long updateId = memberService.adminMemberUpdate(updateForm, updateForm.getMultipartFile());
 
-        Member member = memberService.findOne(updateId);
+        Member member = memberCommonService.findOne(updateId);
         log.info("updated member password : [{}]", member.getPwd());
 
         return "redirect:/admin";
